(BSR)[BO] chore: remove 'WIP_ENABLE_BO_OFFER_DETAILS_V2' feature flag
